Refractor to fit with package standards
