--- tests/prism_translation/resources.txt
@@ -1,5 +1,0 @@
-https://github.com/ahartmanns/jani-models/
-
-https://qcomp.org/benchmarks/index.html
-
-https://momba.dev/guide/construction/
